Refactor helpers module
