+++ generators/app/templates/lib/helpers/get-config.js
@@ -1,4 +1,5 @@
 const config = require('../config');
+const validateEnv = require('./validate-env');
 
 const DEFAULT_ENV = 'stage';
 
@@ -6,18 +7,6 @@ const DEFAULT_ENV = 'stage';
 const getConfig = () => {
   const environment = validateEnv(process.env.NODE_ENV || DEFAULT_ENV);
   return config[environment];
-};
-
-// Ensure tha the custom environment exists. If not throw an ERROR
-const validateEnv = env => {
-  const envAvailable = ['production', 'stage', 'test'];
-  if (!envAvailable.includes(env)) {
-    throw new Error(
-      `\x1b[31mInvalid NODE_ENV on startup: "${env}" is not configured.\n Current available environments: ${envAvailable}\x1b[0m`
-    );
-  } else {
-    return env;
-  }
 };
 
 module.exports = getConfig;
